Slight Changes
Adding a more robust data system less needing to directly choose files and whatnot.
