--- src/main/java/MVC/CheckersBoard.java
@@ -1,4 +1,0 @@
-package MVC;
-
-public class CheckersBoard {
-}
