--- src/main/java/MVC/chModel/Checkers/Pieces/SerfPiece.java
@@ -1,11 +1,0 @@
-package MVC.chModel.Checkers.Pieces;
-
-import MVC.chModel.Checkers.Position;
-
-public class SerfPiece extends AbstractPiece {
-
-
-    public SerfPiece(Position position, PEICE_TEAM team) {
-        super(position, team);
-    }
-}
