--- src/main/java/chNetwork/Client/clientRequest.java
@@ -1,20 +1,0 @@
-package chNetwork.Client;
-
-/***
- * This Class is like super duper important for server things (not really)...
- *
- * it just sends things as simple events.
- * This is going to be the class that
- * kind of helps with that stuff...
- */
-public class clientRequest {
-
-    public clientRequest()
-    {
-
-    }
-
-
-
-
-}
