--- src/main/java/chNetwork/Server/ServerView.java
@@ -1,11 +1,0 @@
-package chNetwork.Server;
-
-import java.util.ArrayList;
-
-//Idk this is important...
-public interface ServerView {
-    void appendMessage(String message);
-    void showErrorMessage(String title, String message);
-    void setWindowTitle(String title);
-    void updateUserList(ArrayList<String> userlist);
-}
