+++ src/main/java/chkMVC/CheckersBoard.java
@@ -1,0 +1,4 @@
+package chkMVC;
+
+public class CheckersBoard {
+}
